refactor: move resource loader package to ext
