+++ ext/resource/loader/loader.go
